# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EchoGuide API
  version: 1.1.0
  description: |
    Contract for the EchoGuide backend. This file is the single source of truth: the Zod
    schemas, TypeScript types, Kotlin client and docs reference are all generated from it.

    **Authentication.** Devices call `register-device` once and receive an opaque session token.
    Every user route requires both `Authorization: Bearer <session_token>` and the
    `X-Install-ID` the token was issued to. Admin routes use a separate admin credential.

    **Correlation.** Every response carries `X-Request-ID`. A client may send its own
    `X-Request-ID` (a UUID); anything else is replaced with a server-generated id.

    **Errors.** Every error body is `{"error": "<safe message>"}`. Provider and database detail
    is never returned.
  license:
    name: Proprietary
    url: https://github.com/dawitlabs/EchoGuide
servers:
  - url: https://api.echoguide.app
    description: Production
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /v1/auth/register-device:
    post:
      tags: [auth]
      operationId: registerDevice
      summary: Register a device and issue a session
      description: |
        A new install id creates an account and a session. Re-registering an install id that is
        already registered requires that device's current session in `Authorization`; the old
        session is revoked and a new one issued. Without it the answer is 409 and the client
        should start a new install. Rate limited per client IP.
      security:
        - {}
        - sessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '200':
          description: Device registered; the session is bound to the install id.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Rotate the session token
      description: |
        Revokes the presented token and issues a new one. A token that expired recently (within
        the server's refresh grace period) may still be refreshed; a revoked token never can.
      parameters:
        - $ref: '#/components/parameters/InstallId'
      responses:
        '200':
          description: The new session. The presented token no longer works.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke the current session
      parameters:
        - $ref: '#/components/parameters/InstallId'
      responses:
        '204':
          description: The session is revoked.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
security:
  - sessionToken: []
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        Opaque session token (`egs_…`) from register-device, bound to the device's install id.
        Must be sent together with the matching `X-Install-ID`. Only a hash is stored server-side.
    adminSession:
      type: http
      scheme: bearer
      description: |
        Separate admin credential (`ega_…`), issued by an operator. Each admin route also requires
        a named permission row. A user session token is never accepted here.
  headers:
    NoStore:
      description: Responses carrying credentials must not be cached.
      schema:
        type: string
        enum: [no-store]
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        minimum: 1
  parameters:
...
    InstallId:
      name: X-Install-ID
      in: header
      required: true
      description: The install id the session token was issued to. Not a credential on its own.
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  responses:
    BadRequest:
      description: The request failed validation, or the audio was shorter than 0.4 s or longer than 15 s.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing, malformed, unknown, expired or revoked credential, or a token presented without its install id.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The account is suspended, or the admin lacks the permission this route requires.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: |
        The request conflicts with current state: the install id is already registered, a request
        with this idempotency key is still running, the term already exists or the vocabulary is
        full, or the account is being deleted.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: The body exceeds the route's limit (small for JSON routes, audio-sized for commands).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A safe, human-readable message.
    Locale:
      type: string
      enum: [am-ET, en-US]
...
    RegisterDeviceRequest:
      type: object
      required: [install_id]
      properties:
        install_id:
          type: string
          minLength: 16
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]{16,128}$'
          description: Generated on the device (a UUID). Identifies the installation, not the person.
        model:
          type: string
          minLength: 1
          maxLength: 128
          description: Device model, for crash and latency analysis.
        locale:
          $ref: '#/components/schemas/Locale'
        phone_hash:
          type: string
          minLength: 1
          maxLength: 128
          deprecated: true
          description: |
            Accepted and ignored. Linking installs by phone number needs a verified number, and no
            verification exists yet; an unverified hash would let anyone claim an account.
    RegisterDeviceResponse:
      type: object
      required: [status, session_token, expires_in]
      properties:
        status:
          type: string
          enum: [registered]
        session_token:
          type: string
          description: Opaque. Revocable server-side; rotated on every refresh.
        expires_in:
          type: integer
          minimum: 1
          description: Seconds until the session expires.
    RefreshSessionResponse:
      type: object
      required: [status, session_token, expires_in]
      properties:
        status:
          type: string
          enum: [refreshed]
        session_token:
          type: string
        expires_in:
          type: integer
          minimum: 1
...